More refactoring for additional simplicity
Removed the two long if/else statements checking each dice roll and then setting the correct image. Found a much shorter way of selecting the correct image based on the random number by using the variables random1 and random2 and inserting them into the filepath, then using that as the image source.